requirements.txt
Django==1.7
Markdown==2.4.1
Pillow==2.5.3
django-bower==5.0.1
django-extensions==1.3.11
django-filter==0.7
djangorestframework==2.3.14
six==1.7.3
wsgiref==0.1.2
